% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits.R
\name{get_cv_indices}
\alias{get_cv_indices}
\title{Get CV indices}
\usage{
get_cv_indices(z, trait)
}
\arguments{
\item{z}{The Z object (loaded from JSON)}

\item{trait}{Name of the trait}
}
\value{
Data frame with 0/1 values for splits
}
\description{
This function retrieves the cross-validation indices for a specific trait from a Z object.
}
