% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{list_species}
\alias{list_species}
\title{List species}
\usage{
list_species(verbose = TRUE, detailed = TRUE)
}
\arguments{
\item{verbose}{If TRUE, prints a nicely formatted list}

\item{detailed}{If TRUE, displays extended metadata in tabular format}
}
\value{
List of species names (sorted, case-preserved)
}
\description{
Lists all available species in the EasyGeSe dataset.
}
