% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits.R
\name{list_traits}
\alias{list_traits}
\title{List traits}
\usage{
list_traits(obj)
}
\arguments{
\item{obj}{Z object (list) or Y object (data frame)}
}
\value{
Vector of trait names
}
\description{
This function retrieves the names of traits from either a Z object (list) or a Y object (data frame).
}
