% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{load_index}
\alias{load_index}
\title{Load index}
\usage{
load_index(force_refresh = FALSE)
}
\arguments{
\item{force_refresh}{If TRUE, force re-download of index}
}
\value{
The loaded index as a list
}
\description{
Loads the index.json file from the EasyGeSe GitHub repository or from local cache.
}
