% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{download_data}
\alias{download_data}
\title{Download data}
\usage{
download_data(species, output_dir = NULL)
}
\arguments{
\item{species}{Species name to download}

\item{output_dir}{Directory to save files. Defaults to user cache directory}
}
\value{
Path to directory containing the downloaded files
}
\description{
Downloads the X (genotype), Y (phenotype), and Z (CV splits) datasets for a specified species.
}
