% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{load_species}
\alias{load_species}
\title{Load species}
\usage{
load_species(species, download = FALSE, download_dir = NULL)
}
\arguments{
\item{species}{Name of the species to load}

\item{download}{If TRUE, download data to local storage}

\item{download_dir}{Directory to save files if downloading. Defaults to user cache directory}
}
\value{
A list containing X, Y, and Z data
}
\description{
Loads the X (genotype), Y (phenotype), and Z (CV splits) datasets for a specified species.
}
